// frontend/src/types/pumps/types.ts
export interface PumpRecord {
  id: number;
  pump_name: string;
  contact_no?: string;
  email_id?: string;
  contact_person?: string;
  address_1?: string;
  address_2?: string;
  address_3?: string;
  opening_balance?: string;
  gstn_no?: string;
  pan_no?: string;
  banker_name?: string;
  branch_name?: string;
  account_no?: string;
  ifsc_code?: string;
  created_at?: string;
  updated_at?: string;
}